Moved cache-related code into subpackage
