+++ sources/src/provision.ts
@@ -8,7 +8,7 @@ import * as toolCache from '@actions/tool-cache'
 
 import * as gradlew from './gradlew'
 import * as params from './input-params'
-import {handleCacheFailure} from './cache-utils'
+import {handleCacheFailure} from './caching/cache-utils'
 import {CacheConfig} from './input-params'
 
 const gradleVersionsBaseUrl = 'https://services.gradle.org/versions'
