+++ sources/test/jest/cache-reporting.test.ts
@@ -1,4 +1,4 @@
-import {CacheEntryListener, CacheListener} from '../../src/cache-reporting'
+import {CacheEntryListener, CacheListener} from '../../src/caching/cache-reporting'
 
 describe('caching report', () => {
     describe('reports not fully restored', () => {
